update to use next/back templatetags and utility class
