+++ templates/decisions/snippets/back_link.html
@@ -1,0 +1,1 @@
+<a href="{{ target }}" id="back">Back</a>
