+++ templates/decisions/snippets/next_button.html
@@ -1,0 +1,3 @@
+<p class="center">
+    <button id="next" type="button"><a href="{{ target }}">{{ btnLabel }}</a></button>
+</p>
