+++ templates/module0/archetype.html
@@ -1,5 +1,6 @@
 {% extends 'decisions/base.html' %}
 {% load static account socialaccount bootstrap i18n %}
+{% load module_tags %}
 {% block title %}{% block h1 %}Module {{ module.display_num }} - Description of Problem Solver
 {% endblock %}{% endblock %}
 {% block content %}
@@ -43,13 +44,10 @@
 
     {% include 'module0/maps/map_2.html' %}
 
-    {% if archetype %}
-        <div>{% with 'module0/archetype/'|add:archetype|add:'.html' as path %}
+    {% if module.archetype %}
+        <div>{% with 'module0/archetype/'|add:module.archetype|add:'.html' as path %}
             {% include path with show_nav=True %}{% endwith %}</div>
     {% endif %}
 
-    <div class="center">
-        <a href="/0/game" id="back">Back</a>
-    </div>
-
+    {% get_back_link module nav.previousUrl %}
 {% endblock %}
