+++ templates/module0/psp_profiles.html
@@ -1,6 +1,8 @@
 {% extends 'decisions/base.html' %}
 {% load static account socialaccount bootstrap i18n %}
-{% block title %}{% block h1 %}Module {{ module.display_num }} - Discover your Problem Solver Profile{% endblock %}{% endblock %}
+{% load module_tags %}
+{% block title %}{% block h1 %}Module {{ module.display_num }} - Discover your Problem Solver Profile
+{% endblock %}{% endblock %}
 {% block content %}
     <style type="text/css">
         h3 img {
@@ -48,33 +50,32 @@
     </style>
 
     <h2>Discover your PROBLEM SOLVER PROFILE</h2>
-    <p class="margin-top-30-mob center margin-0-auto">
-        <h3>5 Problem Solver Profiles</h3>
-	<p>
+
+    <h3>5 Problem Solver Profiles</h3>
+    <p>
         Are you a
         <span class="archetype" title="Albert Einstein">
         <img src="{% static "area_app/img/thinker.png" %}"/>Thinker</span>
         or a
         <span class="archetype" title="Marie Curie">
         <img src="{% static "area_app/img/detective.png" %}"/>Detective</span>?
-	</p>
-	<p>
+    </p>
+    <p>
         Might you solve problems like famous
         <span class="archetype" title="Ernest Shackleton">
         <img src="{% static "area_app/img/adventurer.png" %}"/>Adventurers</span>
         or
         <span class="archetype" title="Oprah Winfrey">
         <img src="{% static "area_app/img/listener.png" %}"/>Listeners</span>?
-	</p>
-	<p>
+    </p>
+    <p>
         Or are you a
         <span class="archetype" title="Beyoncé">
         <img src="{% static "area_app/img/creative.png" %}"/>Visionary</span>
         going your own way?
-	</p>
     </p>
-    <div class="center start">
-        <button id="next" type="button"><a href="/{{ module.num }}/game">Quiz Me</a></button>
-    </div>
-    <a href="/{{ module.num }}/instructions" id="back">Back</a>
+
+    {% get_next_btn module "Quiz Me" nav.nextUrl %}
+
+    {% get_back_link module nav.previousUrl %}
 {% endblock %}
